Keep converting tests to new syntax
