+++ packages/ember/tests/routing/basic_test.js
@@ -3,15 +3,11 @@ import Logger from 'ember-console';
 import {
   Controller,
   RSVP,
-  Object as EmberObject,
-  A as emberA
+  Object as EmberObject
 } from 'ember-runtime';
 import { Route } from 'ember-routing';
 import {
-  run,
-  get,
-  Mixin,
-  observer
+  run
 } from 'ember-metal';
 import {
   Component,
@@ -21,7 +17,6 @@ import {
 import { ENV } from 'ember-environment';
 import { compile } from 'ember-template-compiler';
 import { Application, Engine } from 'ember-application';
-import { Transition } from 'router';
 import { getTextOf } from 'internal-test-helpers';
 
 let Router, App, router, registry, container, originalLoggerError, originalRenderSupport, rootElement;
@@ -89,767 +84,6 @@ QUnit.module('Basic Routing', {
       ENV._ENABLE_RENDER_SUPPORT = originalRenderSupport;
     });
   }
-});
-
-QUnit.test('Generated route should be an instance of App.Route if provided', function(assert) {
-  let generatedRoute;
-
-  Router.map(function() {
-    this.route('posts');
-  });
-
-  App.Route = Route.extend();
-
-  bootApplication();
-
-  handleURL(assert, '/posts');
-
-  generatedRoute = container.lookup('route:posts');
-
-  assert.ok(generatedRoute instanceof App.Route, 'should extend the correct route');
-});
-
-QUnit.test('Nested index route is not overridden by parent\'s implicit index route', function(assert) {
-  Router.map(function() {
-    this.route('posts', function() {
-      this.route('index', { path: ':category' });
-    });
-  });
-
-  bootApplication();
-
-  run(() => router.transitionTo('posts', { category: 'emberjs' }));
-
-  assert.deepEqual(router.location.path, '/posts/emberjs');
-});
-
-QUnit.test('Application template does not duplicate when re-rendered', function(assert) {
-  setTemplate('application', compile('<h3 class="render-once">I render once</h3>{{outlet}}'));
-
-  Router.map(function() {
-    this.route('posts');
-  });
-
-  App.ApplicationRoute = Route.extend({
-    model() {
-      return emberA();
-    }
-  });
-
-  bootApplication();
-
-  // should cause application template to re-render
-  handleURL(assert, '/posts');
-
-  assert.equal(getTextOf(rootElement.querySelector('h3.render-once')), "I render once");
-});
-
-QUnit.test('Child routes should render inside the application template if the application template causes a redirect', function(assert) {
-  setTemplate('application', compile('<h3>App</h3> {{outlet}}'));
-  setTemplate('posts', compile('posts'));
-
-  Router.map(function() {
-    this.route('posts');
-    this.route('photos');
-  });
-
-  App.ApplicationRoute = Route.extend({
-    afterModel() {
-      this.transitionTo('posts');
-    }
-  });
-
-  bootApplication();
-
-  assert.equal(rootElement.textContent.trim(), 'App posts');
-});
-
-QUnit.test('The template is not re-rendered when the route\'s context changes', function(assert) {
-  Router.map(function() {
-    this.route('page', { path: '/page/:name' });
-  });
-
-  App.PageRoute = Route.extend({
-    model(params) {
-      return EmberObject.create({ name: params.name });
-    }
-  });
-
-  let insertionCount = 0;
-  App.FooBarComponent = Component.extend({
-    didInsertElement() {
-      insertionCount += 1;
-    }
-  });
-
-  setTemplate('page', compile(
-    '<p>{{model.name}}{{foo-bar}}</p>'
-  ));
-
-  bootApplication();
-
-  handleURL(assert, '/page/first');
-
-  assert.equal(getTextOf(rootElement.querySelector('p')), 'first');
-  assert.equal(insertionCount, 1);
-
-  handleURL(assert, '/page/second');
-
-  assert.equal(getTextOf(rootElement.querySelector('p')), 'second');
-  assert.equal(insertionCount, 1, 'view should have inserted only once');
-
-  run(() => router.transitionTo('page', EmberObject.create({ name: 'third' })));
-
-  assert.equal(getTextOf(rootElement.querySelector('p')), 'third');
-  assert.equal(insertionCount, 1, 'view should still have inserted only once');
-});
-
-QUnit.test('The template is not re-rendered when two routes present the exact same template & controller', function(assert) {
-  Router.map(function() {
-    this.route('first');
-    this.route('second');
-    this.route('third');
-    this.route('fourth');
-  });
-
-  // Note add a component to test insertion
-
-  let insertionCount = 0;
-  App.XInputComponent = Component.extend({
-    didInsertElement() {
-      insertionCount += 1;
-    }
-  });
-
-  App.SharedRoute = Route.extend({
-    setupController() {
-      this.controllerFor('shared').set('message', 'This is the ' + this.routeName + ' message');
-    },
-
-    renderTemplate() {
-      this.render('shared', { controller: 'shared' });
-    }
-  });
-
-  App.FirstRoute  = App.SharedRoute.extend();
-  App.SecondRoute = App.SharedRoute.extend();
-  App.ThirdRoute  = App.SharedRoute.extend();
-  App.FourthRoute = App.SharedRoute.extend();
-
-  App.SharedController = Controller.extend();
-
-  setTemplate('shared', compile(
-    '<p>{{message}}{{x-input}}</p>'
-  ));
-
-  bootApplication();
-
-  handleURL(assert, '/first');
-
-  assert.equal(getTextOf(rootElement.querySelector('p')), 'This is the first message');
-  assert.equal(insertionCount, 1, 'expected one assertion');
-
-  // Transition by URL
-  handleURL(assert, '/second');
-
-  assert.equal(getTextOf(rootElement.querySelector('p')), 'This is the second message');
-  assert.equal(insertionCount, 1, 'expected one assertion');
-
-  // Then transition directly by route name
-  run(() => {
-    router.transitionTo('third').then(function() {
-      assert.ok(true, 'expected transition');
-    }, function(reason) {
-      assert.ok(false, 'unexpected transition failure: ', QUnit.jsDump.parse(reason));
-    });
-  });
-
-  assert.equal(getTextOf(rootElement.querySelector('p')), 'This is the third message');
-  assert.equal(insertionCount, 1, 'expected one assertion');
-
-  // Lastly transition to a different view, with the same controller and template
-  handleURL(assert, '/fourth');
-  assert.equal(insertionCount, 1, 'expected one assertion');
-
-  assert.equal(getTextOf(rootElement.querySelector('p')), 'This is the fourth message');
-});
-
-QUnit.test('ApplicationRoute with model does not proxy the currentPath', function(assert) {
-  let model = {};
-  let currentPath;
-
-  App.ApplicationRoute = Route.extend({
-    model() { return model; }
-  });
-
-  App.ApplicationController = Controller.extend({
-    currentPathDidChange: observer('currentPath', function() {
-      currentPath = get(this, 'currentPath');
-    })
-  });
-
-  bootApplication();
-
-  assert.equal(currentPath, 'index', 'currentPath is index');
-  assert.equal('currentPath' in model, false, 'should have defined currentPath on controller');
-});
-
-QUnit.test('Promises encountered on app load put app into loading state until resolved', function(assert) {
-  assert.expect(2);
-
-  let deferred = RSVP.defer();
-
-  App.IndexRoute = Route.extend({
-    model() {
-      return deferred.promise;
-    }
-  });
-
-  setTemplate('index', compile('<p>INDEX</p>'));
-  setTemplate('loading', compile('<p>LOADING</p>'));
-
-  bootApplication();
-
-  assert.equal(getTextOf(rootElement.querySelector('p')), 'LOADING', 'The loading state is displaying.');
-  run(deferred.resolve);
-  assert.equal(getTextOf(rootElement.querySelector('p')), 'INDEX', 'The index route is display.');
-});
-
-QUnit.test('Route should tear down multiple outlets', function(assert) {
-  setTemplate('application', compile('{{outlet \'menu\'}}{{outlet}}{{outlet \'footer\'}}'));
-  setTemplate('posts', compile('{{outlet}}'));
-  setTemplate('users', compile('users'));
-  setTemplate('posts/index', compile('<p class="posts-index">postsIndex</p>'));
-  setTemplate('posts/menu', compile('<div class="posts-menu">postsMenu</div>'));
-  setTemplate('posts/footer', compile('<div class="posts-footer">postsFooter</div>'));
-
-  Router.map(function() {
-    this.route('posts', function() {});
-    this.route('users', function() {});
-  });
-
-  App.PostsRoute = Route.extend({
-    renderTemplate() {
-      this.render('posts/menu', {
-        into: 'application',
-        outlet: 'menu'
-      });
-
-      this.render();
-
-      this.render('posts/footer', {
-        into: 'application',
-        outlet: 'footer'
-      });
-    }
-  });
-
-  bootApplication();
-
-  handleURL(assert, '/posts');
-
-  assert.equal(getTextOf(rootElement.querySelector('div.posts-menu')), 'postsMenu', 'The posts/menu template was rendered');
-  assert.equal(getTextOf(rootElement.querySelector('p.posts-index')), 'postsIndex', 'The posts/index template was rendered');
-  assert.equal(getTextOf(rootElement.querySelector('div.posts-footer')), 'postsFooter', 'The posts/footer template was rendered');
-
-  handleURL(assert, '/users');
-
-  assert.equal(rootElement.querySelector('div.posts-menu'), null, 'The posts/menu template was removed');
-  assert.equal(rootElement.querySelector('p.posts-index'), null, 'The posts/index template was removed');
-  assert.equal(rootElement.querySelector('div.posts-footer'), null, 'The posts/footer template was removed');
-});
-
-QUnit.test('Route will assert if you try to explicitly render {into: ...} a missing template', function() {
-  expectDeprecation(/Rendering into a {{render}} helper that resolves to an {{outlet}} is deprecated./);
-
-  Router.map(function() {
-    this.route('home', { path: '/' });
-  });
-
-  App.HomeRoute = Route.extend({
-    renderTemplate() {
-      this.render({ into: 'nonexistent' });
-    }
-  });
-
-  expectAssertion(() => bootApplication(), 'You attempted to render into \'nonexistent\' but it was not found');
-});
-
-QUnit.test('Route supports clearing outlet explicitly', function(assert) {
-  setTemplate('application', compile('{{outlet}}{{outlet \'modal\'}}'));
-  setTemplate('posts', compile('{{outlet}}'));
-  setTemplate('users', compile('users'));
-  setTemplate('posts/index', compile('<div class="posts-index">postsIndex {{outlet}}</div>'));
-  setTemplate('posts/modal', compile('<div class="posts-modal">postsModal</div>'));
-  setTemplate('posts/extra', compile('<div class="posts-extra">postsExtra</div>'));
-
-  Router.map(function() {
-    this.route('posts', function() {});
-    this.route('users', function() {});
-  });
-
-  App.PostsRoute = Route.extend({
-    actions: {
-      showModal() {
-        this.render('posts/modal', {
-          into: 'application',
-          outlet: 'modal'
-        });
-      },
-      hideModal() {
-        this.disconnectOutlet({ outlet: 'modal', parentView: 'application' });
-      }
-    }
-  });
-
-  App.PostsIndexRoute = Route.extend({
-    actions: {
-      showExtra() {
-        this.render('posts/extra', {
-          into: 'posts/index'
-        });
-      },
-      hideExtra() {
-        this.disconnectOutlet({ parentView: 'posts/index' });
-      }
-    }
-  });
-
-  bootApplication();
-
-  handleURL(assert, '/posts');
-
-  assert.equal(getTextOf(rootElement.querySelector('div.posts-index')), 'postsIndex', 'The posts/index template was rendered');
-
-  run(() => router.send('showModal'));
-
-  assert.equal(getTextOf(rootElement.querySelector('div.posts-modal')), 'postsModal', 'The posts/modal template was rendered');
-
-  run(() => router.send('showExtra'));
-
-  assert.equal(getTextOf(rootElement.querySelector('div.posts-extra')), 'postsExtra', 'The posts/extra template was rendered');
-
-  run(() => router.send('hideModal'));
-
-  assert.equal(rootElement.querySelector('div.posts-modal'), null, 'The posts/modal template was removed');
-
-  run(() => router.send('hideExtra'));
-
-  assert.equal(rootElement.querySelector('div.posts-extra'), null, 'The posts/extra template was removed');
-  run(function() {
-    router.send('showModal');
-  });
-  assert.equal(getTextOf(rootElement.querySelector('div.posts-modal')), 'postsModal', 'The posts/modal template was rendered');
-  run(function() {
-    router.send('showExtra');
-  });
-  assert.equal(getTextOf(rootElement.querySelector('div.posts-extra')), 'postsExtra', 'The posts/extra template was rendered');
-
-  handleURL(assert, '/users');
-
-  assert.equal(rootElement.querySelector('div.posts-index'), null, 'The posts/index template was removed');
-  assert.equal(rootElement.querySelector('div.posts-modal'), null, 'The posts/modal template was removed');
-  assert.equal(rootElement.querySelector('div.posts-extra'), null, 'The posts/extra template was removed');
-});
-
-QUnit.test('Route supports clearing outlet using string parameter', function(assert) {
-  setTemplate('application', compile('{{outlet}}{{outlet \'modal\'}}'));
-  setTemplate('posts', compile('{{outlet}}'));
-  setTemplate('users', compile('users'));
-  setTemplate('posts/index', compile('<div class="posts-index">postsIndex {{outlet}}</div>'));
-  setTemplate('posts/modal', compile('<div class="posts-modal">postsModal</div>'));
-
-  Router.map(function() {
-    this.route('posts', function() {});
-    this.route('users', function() {});
-  });
-
-  App.PostsRoute = Route.extend({
-    actions: {
-      showModal() {
-        this.render('posts/modal', {
-          into: 'application',
-          outlet: 'modal'
-        });
-      },
-      hideModal() {
-        this.disconnectOutlet('modal');
-      }
-    }
-  });
-
-  bootApplication();
-
-  handleURL(assert, '/posts');
-
-  assert.equal(getTextOf(rootElement.querySelector('div.posts-index')), 'postsIndex', 'The posts/index template was rendered');
-
-  run(() => router.send('showModal'));
-
-  assert.equal(getTextOf(rootElement.querySelector('div.posts-modal')), 'postsModal', 'The posts/modal template was rendered');
-
-  run(() => router.send('hideModal'));
-
-  assert.equal(rootElement.querySelector('div.posts-modal'), null, 'The posts/modal template was removed');
-
-  handleURL(assert, '/users');
-
-  assert.equal(rootElement.querySelector('div.posts-index'), null, 'The posts/index template was removed');
-  assert.equal(rootElement.querySelector('div.posts-modal'), null, 'The posts/modal template was removed');
-});
-
-QUnit.test('Route silently fails when cleaning an outlet from an inactive view', function(assert) {
-  assert.expect(1); // handleURL
-
-  setTemplate('application', compile('{{outlet}}'));
-  setTemplate('posts', compile('{{outlet \'modal\'}}'));
-  setTemplate('modal', compile('A Yo.'));
-
-  Router.map(function() {
-    this.route('posts');
-  });
-
-  App.PostsRoute = Route.extend({
-    actions: {
-      hideSelf() {
-        this.disconnectOutlet({ outlet: 'main', parentView: 'application' });
-      },
-      showModal() {
-        this.render('modal', { into: 'posts', outlet: 'modal' });
-      },
-      hideModal() {
-        this.disconnectOutlet({ outlet: 'modal', parentView: 'posts' });
-      }
-    }
-  });
-
-  bootApplication();
-
-  handleURL(assert, '/posts');
-
-  run(() => router.send('showModal'));
-  run(() => router.send('hideSelf'));
-  run(() => router.send('hideModal'));
-});
-
-QUnit.test('Router `willTransition` hook passes in cancellable transition', function(assert) {
-  // Should hit willTransition 3 times, once for the initial route, and then 2 more times
-  // for the two handleURL calls below
-  assert.expect(3);
-
-  Router.map(function() {
-    this.route('nork');
-    this.route('about');
-  });
-
-  Router.reopen({
-    init() {
-      this._super();
-      this.on('willTransition', this.testWillTransitionHook);
-    },
-    testWillTransitionHook(transition, url) {
-      assert.ok(true, 'willTransition was called ' + url);
-      transition.abort();
-    }
-  });
-
-  App.LoadingRoute = Route.extend({
-    activate() {
-      assert.ok(false, 'LoadingRoute was not entered');
-    }
-  });
-
-  App.NorkRoute = Route.extend({
-    activate() {
-      assert.ok(false, 'NorkRoute was not entered');
-    }
-  });
-
-  App.AboutRoute = Route.extend({
-    activate() {
-      assert.ok(false, 'AboutRoute was not entered');
-    }
-  });
-
-  bootApplication();
-
-  // Attempted transitions out of index should abort.
-  run(router, 'handleURL', '/nork');
-  run(router, 'handleURL', '/about');
-});
-
-QUnit.test('Aborting/redirecting the transition in `willTransition` prevents LoadingRoute from being entered', function(assert) {
-  assert.expect(8);
-
-  Router.map(function() {
-    this.route('nork');
-    this.route('about');
-  });
-
-  let redirect = false;
-
-  App.IndexRoute = Route.extend({
-    actions: {
-      willTransition(transition) {
-        assert.ok(true, 'willTransition was called');
-        if (redirect) {
-          // router.js won't refire `willTransition` for this redirect
-          this.transitionTo('about');
-        } else {
-          transition.abort();
-        }
-      }
-    }
-  });
-
-  let deferred = null;
-
-  App.LoadingRoute = Route.extend({
-    activate() {
-      assert.ok(deferred, 'LoadingRoute should be entered at this time');
-    },
-    deactivate() {
-      assert.ok(true, 'LoadingRoute was exited');
-    }
-  });
-
-  App.NorkRoute = Route.extend({
-    activate() {
-      assert.ok(true, 'NorkRoute was entered');
-    }
-  });
-
-  App.AboutRoute = Route.extend({
-    activate() {
-      assert.ok(true, 'AboutRoute was entered');
-    },
-    model() {
-      if (deferred) { return deferred.promise; }
-    }
-  });
-
-  bootApplication();
-
-  // Attempted transitions out of index should abort.
-  run(router, 'transitionTo', 'nork');
-  run(router, 'handleURL', '/nork');
-
-  // Attempted transitions out of index should redirect to about
-  redirect = true;
-  run(router, 'transitionTo', 'nork');
-  run(router, 'transitionTo', 'index');
-
-  // Redirected transitions out of index to a route with a
-  // promise model should pause the transition and
-  // activate LoadingRoute
-  deferred = RSVP.defer();
-  run(router, 'transitionTo', 'nork');
-  run(deferred.resolve);
-});
-
-QUnit.test('`didTransition` event fires on the router', function(assert) {
-  assert.expect(3);
-
-  Router.map(function() {
-    this.route('nork');
-  });
-
-  router = container.lookup('router:main');
-
-  router.one('didTransition', function() {
-    assert.ok(true, 'didTransition fired on initial routing');
-  });
-
-  bootApplication();
-
-  router.one('didTransition', function() {
-    assert.ok(true, 'didTransition fired on the router');
-    assert.equal(router.get('url'), '/nork', 'The url property is updated by the time didTransition fires');
-  });
-
-  run(router, 'transitionTo', 'nork');
-});
-QUnit.test('`didTransition` can be reopened', function(assert) {
-  assert.expect(1);
-
-  Router.map(function() {
-    this.route('nork');
-  });
-
-  Router.reopen({
-    didTransition() {
-      this._super(...arguments);
-      assert.ok(true, 'reopened didTransition was called');
-    }
-  });
-
-  bootApplication();
-});
-
-QUnit.test('`activate` event fires on the route', function(assert) {
-  assert.expect(2);
-
-  let eventFired = 0;
-
-  Router.map(function() {
-    this.route('nork');
-  });
-
-  App.NorkRoute = Route.extend({
-    init() {
-      this._super(...arguments);
-
-      this.on('activate', function() {
-        assert.equal(++eventFired, 1, 'activate event is fired once');
-      });
-    },
-
-    activate() {
-      assert.ok(true, 'activate hook is called');
-    }
-  });
-
-  bootApplication();
-
-  run(router, 'transitionTo', 'nork');
-});
-
-QUnit.test('`deactivate` event fires on the route', function(assert) {
-  assert.expect(2);
-
-  let eventFired = 0;
-
-  Router.map(function() {
-    this.route('nork');
-    this.route('dork');
-  });
-
-  App.NorkRoute = Route.extend({
-    init() {
-      this._super(...arguments);
-
-      this.on('deactivate', function() {
-        assert.equal(++eventFired, 1, 'deactivate event is fired once');
-      });
-    },
-
-    deactivate() {
-      assert.ok(true, 'deactivate hook is called');
-    }
-  });
-
-  bootApplication();
-
-  run(router, 'transitionTo', 'nork');
-  run(router, 'transitionTo', 'dork');
-});
-
-QUnit.test('Actions can be handled by inherited action handlers', function(assert) {
-  assert.expect(4);
-
-  App.SuperRoute = Route.extend({
-    actions: {
-      foo() {
-        assert.ok(true, 'foo');
-      },
-      bar(msg) {
-        assert.equal(msg, 'HELLO');
-      }
-    }
-  });
-
-  App.RouteMixin = Mixin.create({
-    actions: {
-      bar(msg) {
-        assert.equal(msg, 'HELLO');
-        this._super(msg);
-      }
-    }
-  });
-
-  App.IndexRoute = App.SuperRoute.extend(App.RouteMixin, {
-    actions: {
-      baz() {
-        assert.ok(true, 'baz');
-      }
-    }
-  });
-
-  bootApplication();
-
-  router.send('foo');
-  router.send('bar', 'HELLO');
-  router.send('baz');
-});
-
-QUnit.test('transitionTo returns Transition when passed a route name', function(assert) {
-  assert.expect(1);
-  Router.map(function() {
-    this.route('root', { path: '/' });
-    this.route('bar');
-  });
-
-  bootApplication();
-
-  let transition = run(() => router.transitionTo('bar'));
-
-  assert.equal(transition instanceof Transition, true);
-});
-
-QUnit.test('transitionTo returns Transition when passed a url', function(assert) {
-  assert.expect(1);
-  Router.map(function() {
-    this.route('root', { path: '/' });
-    this.route('bar', function() {
-      this.route('baz');
-    });
-  });
-
-  bootApplication();
-
-  let transition = run(() => router.transitionTo('/bar/baz'));
-
-  assert.equal(transition instanceof Transition, true);
-});
-
-QUnit.test('currentRouteName is a property installed on ApplicationController that can be used in transitionTo', function(assert) {
-  assert.expect(24);
-
-  Router.map(function() {
-    this.route('be', function() {
-      this.route('excellent', { resetNamespace: true }, function() {
-        this.route('to', { resetNamespace: true }, function() {
-          this.route('each', { resetNamespace: true }, function() {
-            this.route('other');
-          });
-        });
-      });
-    });
-  });
-
-  bootApplication();
-
-  let appController = getOwner(router).lookup('controller:application');
-
-  function transitionAndCheck(path, expectedPath, expectedRouteName) {
-    if (path) { run(router, 'transitionTo', path); }
-    assert.equal(appController.get('currentPath'), expectedPath);
-    assert.equal(appController.get('currentRouteName'), expectedRouteName);
-  }
-
-  transitionAndCheck(null, 'index', 'index');
-  transitionAndCheck('/be', 'be.index', 'be.index');
-  transitionAndCheck('/be/excellent', 'be.excellent.index', 'excellent.index');
-  transitionAndCheck('/be/excellent/to', 'be.excellent.to.index', 'to.index');
-  transitionAndCheck('/be/excellent/to/each', 'be.excellent.to.each.index', 'each.index');
-  transitionAndCheck('/be/excellent/to/each/other', 'be.excellent.to.each.other', 'each.other');
-
-  transitionAndCheck('index', 'index', 'index');
-  transitionAndCheck('be', 'be.index', 'be.index');
-  transitionAndCheck('excellent', 'be.excellent.index', 'excellent.index');
-  transitionAndCheck('to.index', 'be.excellent.to.index', 'to.index');
-  transitionAndCheck('each', 'be.excellent.to.each.index', 'each.index');
-  transitionAndCheck('each.other', 'be.excellent.to.each.other', 'each.other');
 });
 
 QUnit.test('Route model hook finds the same model as a manual find', function(assert) {
@@ -1896,4 +1130,22 @@ QUnit.test('App.destroy does not leave undestroyed views after clearing engines'
 
   run(App, 'destroy');
   assert.equal(router._toplevelView, null, 'the toplevelView was not reinitialized');
+});
+
+QUnit.test('Generated route should be an instance of App.Route if provided', function (assert) {
+  let generatedRoute;
+
+  Router.map(function () {
+    this.route('posts');
+  });
+
+  App.Route = Route.extend();
+
+  bootApplication();
+
+  handleURL(assert, '/posts');
+
+  generatedRoute = container.lookup('route:posts');
+
+  assert.ok(generatedRoute instanceof App.Route, 'should extend the correct route');
 });
